refactor: preset 변경이 ui에 바로 반영되도록 파일구조 변경
