+++ packages/panda-preset/src/preset.ts
@@ -3,20 +3,21 @@ import { colors, radii, fontSizes, textStyles } from "./tokens/tokens"
 import { recipes, slotRecipes } from "./recipes"
 import { keyframes } from "./tokens/keyframes"
 
-export const preset = definePreset({
-  theme: {
-    extend: {
-      tokens: {
-        radii,
-        fontSizes,
+export const preset = () =>
+  definePreset({
+    theme: {
+      extend: {
+        tokens: {
+          radii,
+          fontSizes,
+        },
+        semanticTokens: {
+          colors,
+        },
+        textStyles,
+        recipes,
+        slotRecipes,
+        keyframes,
       },
-      semanticTokens: {
-        colors,
-      },
-      textStyles,
-      recipes,
-      slotRecipes,
-      keyframes,
     },
-  },
-})
+  })
