+++ packages/ui/package.json
@@ -3,34 +3,13 @@
   "version": "1.5.21",
   "description": "",
   "type": "module",
-  "main": "dist/index.umd.js",
-  "module": "dist/index.es.js",
-  "types": "dist/index.d.ts",
-  "exports": {
-    "./styles.css": "./dist/styles.css",
-    ".": {
-      "types": "./dist/src/components/index.d.ts",
-      "require": "./dist/index.umd.js",
-      "import": "./dist/index.es.js"
-    }
-  },
-  "files": [
-    "dist/",
-    "package.json"
-  ],
   "scripts": {
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "local": "vite",
-    "type": "tsc -p tsconfig.json --skipLibCheck && tsc-alias -p tsconfig.json",
+    "prepare": "panda codegen",
+    "test-storybook": "test-storybook",
     "storybook": "storybook dev -p 6006",
     "build": "pnpm prepare && vite build",
     "build-storybook": "storybook build",
-    "test-storybook": "test-storybook",
-    "dev": "pnpm panda --watch",
-    "codegen": "pnpm panda codegen --clean",
-    "cssgen": "pnpm panda cssgen",
-    "release": "pnpm panda prepare && vite build && pnpm panda cssgen && pnpm build-storybook",
-    "prepare": "pnpm panda codegen --clean",
     "check-types": "tsc --noEmit",
     "studio": "pnpm panda studio",
     "pub": "pnpm publish --no-git-checks"
@@ -45,29 +24,29 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "^1.5.0",
+    "@chromatic-com/storybook": "^3.1.0",
     "@emotion/is-prop-valid": "^1.3.0",
     "@pandacss/dev": "^0.40.1",
     "@pandacss/studio": "^0.46.0",
-    "@storybook/addon-a11y": "^8.1.11",
-    "@storybook/addon-controls": "^8.1.6",
-    "@storybook/addon-docs": "^8.2.2",
-    "@storybook/addon-essentials": "^8.1.5",
-    "@storybook/addon-interactions": "^8.1.11",
-    "@storybook/addon-links": "^8.1.5",
-    "@storybook/addon-onboarding": "^8.1.5",
-    "@storybook/blocks": "^8.1.5",
-    "@storybook/preview-api": "^8.1.5",
-    "@storybook/react": "^8.1.5",
-    "@storybook/react-vite": "^8.2.5",
-    "@storybook/test": "^8.1.11",
+    "@storybook/addon-a11y": "^8.3.6",
+    "@storybook/addon-controls": "^8.3.6",
+    "@storybook/addon-docs": "^8.3.6",
+    "@storybook/addon-essentials": "^8.3.6",
+    "@storybook/addon-interactions": "^8.3.6",
+    "@storybook/addon-links": "^8.3.6",
+    "@storybook/addon-onboarding": "^8.3.6",
+    "@storybook/blocks": "^8.3.6",
+    "@storybook/preview-api": "^8.3.6",
+    "@storybook/react": "^8.3.6",
+    "@storybook/react-vite": "^8.3.6",
+    "@storybook/test": "^8.3.6",
     "@storybook/test-runner": "^0.19.1",
     "@types/is-hotkey": "^0.1.10",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "axe-playwright": "^2.0.1",
     "chromatic": "^11.5.4",
-    "storybook": "^8.1.5",
+    "storybook": "^8.3.6",
     "tsc-alias": "^1.8.10",
     "tslib": "^2.6.3",
     "typescript": "^5.2.2",
