+++ packages/ui/postcss.config.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  plugins: {
+    "@pandacss/dev/postcss": {},
+  },
+}
