+++ packages/ui/src/components/index.ts
@@ -1,5 +1,5 @@
-export * from "./Button/Button"
+export * from "./Button"
 // export * from "./TagButton/TagButton"
-export * from "./Accordion/Accordion"
+export * from "./Accordion"
 // export * from "./Tab/index"
 // export * from "./Slider/Slider"
