--- packages/ui/src/postcss.config.cjs
@@ -1,4 +1,0 @@
-module.exports = {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  plugins: [require("@pandacss/dev/postcss")()],
-}
