+++ packages/ui/tsconfig.json
@@ -14,6 +14,6 @@
       "@styled-system/*": ["styled-system/*"]
     }
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "postcss.config.cjs"],
   "exclude": ["node_modules", "dist"]
 }
